\name{PgsRankRnnotatR-package}
\alias{PgsRankRnnotatR-package}
\alias{PgsRankRnnotatR}
\docType{package}
\title{
\packageTitle{PgsRankRnnotatR}
}
\description{
\packageDescription{PgsRankRnnotatR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PgsRankRnnotatR}
\packageIndices{PgsRankRnnotatR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PgsRankRnnotatR}

Maintainer: \packageMaintainer{PgsRankRnnotatR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
